+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add property `title` to Histogram
 - Add property `title` to Line-Chart
 - Add property `title` to Scatter-Chart
+- Add property `title` to Pie-Chart
 
 ## [0.2.2] - 2019-10-06